adding the function for withdraw and deposit

+++ Server/entities/transactionsEntity.sql
@@ -2,7 +2,7 @@ use bank;
 
 CREATE TABLE transactions(
     transaction_id int PRIMARY KEY AUTO_INCREMENT ,
-    transiction_amount INT ,   
+    transiction_amount FLOAT ,   
     category_name VARCHAR(20),    
     transaction_vendor VARCHAR(40),
     is_delete BIT,
